Implement cross fades
The cross fades are implemented so that they can be added and removed for pairs
of clips without changing the frames rendered in any clip, the clips are
extended into each other from the in and out points.

For now, pressing F toggles a 23 frames long cross fade-in for the current clip.